Main added and code cleanup + quick fixes
