--- src/main/java/knnclassifier/DistanceMeter.java
@@ -1,14 +1,0 @@
-
-package tletters.knnclassifier;
-
-import java.util.Collection;
-
-/**
- *
- * @author Piotr
- * @param <VAL>
- */
-public interface  DistanceMeter <VAL extends Number>
-{
-	public  double distance( Collection<VAL> lhs, Collection<VAL> rhs );
-}
