+++ src/main/java/tletters/featureextraction/Zoning.java
@@ -1,4 +1,4 @@
-package FeatureExtraction;
+package tletters.featureextraction;
 
 import java.awt.image.BufferedImage;
 
@@ -10,6 +10,7 @@ import static java.lang.StrictMath.pow;
  * na podstawie http://www.ijcaonline.org/volume27/number4/pxc3874481.pdf
  */
 public class Zoning implements ExtractionAlgorithm {
+
     public static final int DEF_COLUMNS = 4;
     public static final int DEF_ROWS = 4;
     public static final int DEF_ALPHA = 1;
@@ -17,48 +18,46 @@ public class Zoning implements ExtractionAlgorithm {
     private final int rows;
     private final double alpha;
 
-    public Zoning(){
+    public Zoning() {
         columns = DEF_COLUMNS;
         rows = DEF_ROWS;
         alpha = DEF_ALPHA;
     }
 
-    public Zoning(int rows, int columns, double alpha){
+    public Zoning(int rows, int columns, double alpha) {
         this.columns = columns;
         this.rows = rows;
         this.alpha = alpha;
     }
 
-
     @Override
     public double[] extractFeatures(BufferedImage bufferedImage) {
-        double[] res = new double[ columns * rows];
-        for(int i = 0; i < rows; i++){
-            for(int j =  0; j < columns; j++){
-                res[i * columns + j] = computeCellValue(i,j, bufferedImage);
+        double[] res = new double[columns * rows];
+        for (int i = 0; i < rows; i++) {
+            for (int j = 0; j < columns; j++) {
+                res[i * columns + j] = computeCellValue(i, j, bufferedImage);
             }
         }
         return res;
     }
 
     private Double computeCellValue(int row, int column, BufferedImage bufferedImage) {
-        Double cellSum  = 0.0;
-        int cellWidth = (int) Math.ceil(bufferedImage.getWidth()/columns);
-        int cellHeight = (int) Math.ceil(bufferedImage.getHeight()/rows); // TODO sprwdzic czy nie ma zaokroglania w dol
+        Double cellSum = 0.0;
+        int cellWidth = (int) Math.ceil(bufferedImage.getWidth() / columns);
+        int cellHeight = (int) Math.ceil(bufferedImage.getHeight() / rows); // TODO sprwdzic czy nie ma zaokroglania w dol
         int rowBeg = row * cellHeight;
         int colBeg = column * cellWidth;
-        for(int i = rowBeg; i < Math.min((row + 1) * cellHeight, bufferedImage.getHeight()); i++){
-            for(int j = colBeg; j < Math.min((column + 1) * cellWidth, bufferedImage.getWidth()); j++){
-                if(isBlack(i,j,bufferedImage)){
-                    cellSum+= sqrt( pow((i - rowBeg),2) + pow((j - colBeg),2)) * 1.0/sqrt(1+ pow(((j-colBeg + 1)/(i-rowBeg + 1)),2));
+        for (int i = rowBeg; i < Math.min((row + 1) * cellHeight, bufferedImage.getHeight()); i++) {
+            for (int j = colBeg; j < Math.min((column + 1) * cellWidth, bufferedImage.getWidth()); j++) {
+                if (isBlack(i, j, bufferedImage)) {
+                    cellSum += sqrt(pow((i - rowBeg), 2) + pow((j - colBeg), 2)) * 1.0 / sqrt(1 + pow(((j - colBeg + 1) / (i - rowBeg + 1)), 2));
                     // cos(arctan(x) zredukowany do 1/(sqrt (1+x^2))
                     // w celu uniknięcia dzielenia przez zero dla wyliczania theta użyłęm (y+1)/(x+1)
                     // zamiast y/x nie powinno byc problemow
                 }
             }
         }
-
-        return alpha * cellSum/getNumberOfPixels(row, column, bufferedImage, cellWidth, cellHeight);
+        return alpha * cellSum / getNumberOfPixels(row, column, bufferedImage, cellWidth, cellHeight);
     }
 
     private int getNumberOfPixels(int row, int column, BufferedImage bufferedImage, int cellWidth, int cellHeight) {
@@ -67,7 +66,7 @@ public class Zoning implements ExtractionAlgorithm {
     }
 
     private boolean isBlack(int i, int j, BufferedImage bufferedImage) {
-        return bufferedImage.getRGB(j,i)!= -1;
+        return bufferedImage.getRGB(i, j) <= -16350000 && bufferedImage.getRGB(i, j) > -17000000;
     }
 
 }
