+++ src/test/java/tletters/glyph/GlyphTest.java
@@ -1,5 +1,5 @@
-import glyph.Glyph;
-import glyph.LanguageType;
+package tletters.glyph;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -13,27 +13,27 @@ public class GlyphTest {
     public void testWillThrowExceptionWhenGlyphCaseDoesNotMatchCharacterCase() {
         expectedEx.expect(IllegalArgumentException.class);
         expectedEx.expectMessage("Character case and glyphCase do not match");
-        new Glyph(new double[0], Glyph.GlyphCase.UPPER, Glyph.GlyphType.LETTER, LanguageType.GENERAL_PL, 'c');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, 'c');
     }
 
     @Test
     public void testWillThrowExceptionWhenGlyphTypeDoesNotMatchCharacterType() {
         expectedEx.expect(IllegalArgumentException.class);
         expectedEx.expectMessage("Character type and glyphType do not match");
-        new Glyph(new double[0], Glyph.GlyphCase.UPPER, Glyph.GlyphType.LETTER, LanguageType.GENERAL_PL, '1');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, '1');
     }
 
     @Test
     public void testWillThrowExceptionWhenGlyphTypeIsDigitAndGlyphCaseIsLower() {
         expectedEx.expect(IllegalArgumentException.class);
         expectedEx.expectMessage("Character case and glyphCase do not match");
-        new Glyph(new double[0], Glyph.GlyphCase.LOWER, Glyph.GlyphType.DIGIT, LanguageType.GENERAL_PL, '1');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, '1');
     }
 
     @Test
     public void testCanBeConstructedWithoutExceptionsFoValidTypeCaseCombination() {
-        new Glyph(new double[0], Glyph.GlyphCase.UPPER, Glyph.GlyphType.DIGIT, LanguageType.GENERAL_PL, '7');
-        new Glyph(new double[0], Glyph.GlyphCase.LOWER, Glyph.GlyphType.LETTER, LanguageType.GENERAL_PL, 'a');
-        new Glyph(new double[0], Glyph.GlyphCase.UPPER, Glyph.GlyphType.LETTER, LanguageType.GENERAL_PL, 'A');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, '7');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, 'a');
+        new Glyph(new double[0], LanguageType.GENERAL_PL, 'A');
     }
 }
